Added option to keep intermediates (#28)
* starting, moving to HPC

* Updated completed_preprocess.tsv at 2022-03-15 11:58:48

+++ cli/afni_resting_subj.py
@@ -58,6 +58,7 @@ def submit_jobs(
     do_regress,
     coord_dict,
     do_blur,
+    kp_interm,
 ):
     """Schedule work for single participant.
 
@@ -89,6 +90,8 @@ def submit_jobs(
         seed name and coordinates
     do_blur : bool
         [T/F] whether to blur as part of pre-processing
+    kp_interm : bool
+        [T/F] whether to keep (T) or remove (F) intemediates
 
     Returns
     -------
@@ -137,47 +140,49 @@ def submit_jobs(
                 "{subj}",
                 "{sess}",
                 {coord_dict},
+                {kp_interm},
             )
         print(f"Finished {subj}/{sess}/{task} with: \\n {{afni_data}}")
 
-        # clean up niftis
-        shutil.rmtree(os.path.join("{afni_dir}", "{subj}", "{sess}", "sbatch_out"))
-        clean_dir = os.path.join("{afni_dir}", "{subj}", "{sess}")
-        clean_list = [
-            "preproc_bold",
-            "smoothed_bold",
-            "nuissance_bold",
-            "probseg",
-            "preproc_T1w",
-            "minval_mask",
-            "GMe_mask",
-            "meanTS_bold",
-            "sdTS_bold",
-            "blurWM_bold",
-            "combWM_bold",
-            "masked_bold",
-        ]
-        for c_str in clean_list:
-            for h_file in glob.glob(f"{{clean_dir}}/**/*{{c_str}}.nii.gz", recursive=True):
-                os.remove(h_file)
+        # clean up
+        if not {kp_interm}:
+            shutil.rmtree(os.path.join("{afni_dir}", "{subj}", "{sess}", "sbatch_out"))
+            clean_dir = os.path.join("{afni_dir}", "{subj}", "{sess}")
+            clean_list = [
+                "preproc_bold",
+                "smoothed_bold",
+                "nuissance_bold",
+                "probseg",
+                "preproc_T1w",
+                "minval_mask",
+                "GMe_mask",
+                "meanTS_bold",
+                "sdTS_bold",
+                "blurWM_bold",
+                "combWM_bold",
+                "masked_bold",
+            ]
+            for c_str in clean_list:
+                for h_file in glob.glob(f"{{clean_dir}}/**/*{{c_str}}.nii.gz", recursive=True):
+                    os.remove(h_file)
 
-        # clean up other, based on extension
-        clean_list = [
-            "unit+tlrc.HEAD",
-            "unit+tlrc.BRIK",
-            "corr+tlrc.HEAD",
-            "corr+tlrc.BRIK",
-            "1D00.1D",
-            "1D01.1D",
-            "1D02.1D",
-            "1D_eig.1D",
-            "1D_vec.1D",
-            "csfPC_timeseries.1D",
-            "tmp-censor_timeseries.1D",
-        ]
-        for c_str in clean_list:
-            for h_file in glob.glob(f"{{clean_dir}}/**/*{{c_str}}", recursive=True):
-                os.remove(h_file)
+            # clean up other, based on extension
+            clean_list = [
+                "unit+tlrc.HEAD",
+                "unit+tlrc.BRIK",
+                "corr+tlrc.HEAD",
+                "corr+tlrc.BRIK",
+                "1D00.1D",
+                "1D01.1D",
+                "1D02.1D",
+                "1D_eig.1D",
+                "1D_vec.1D",
+                "csfPC_timeseries.1D",
+                "tmp-censor_timeseries.1D",
+            ]
+            for c_str in clean_list:
+                for h_file in glob.glob(f"{{clean_dir}}/**/*{{c_str}}", recursive=True):
+                    os.remove(h_file)
 
         # copy important files to /home/data
         h_cmd = f"cp -r {afni_dir}/{subj} {afni_final}"
@@ -281,6 +286,16 @@ def get_args():
             """
         ),
     )
+    parser.add_argument(
+        "--keep-interm",
+        action="store_true",
+        help=textwrap.dedent(
+            """\
+            Toggle of whether to remove intermediates.
+            Boolean (True if "--keep-interm", else False).
+            """
+        ),
+    )
 
     required_args = parser.add_argument_group("Required Arguments")
     required_args.add_argument(
@@ -315,6 +330,7 @@ def main():
     task = args.task
     code_dir = args.code_dir
     do_blur = args.blur
+    kp_interm = args.keep_interm
 
     # set up
     # TODO get coord_dict from user-specified JSON
@@ -389,6 +405,7 @@ def main():
             value_dict["Regress"],
             coord_dict,
             do_blur,
+            kp_interm,
         )
         time.sleep(3)
         print(f"submit_jobs out: {h_out} \nsubmit_jobs err: {h_err}")
